[feat/#318] 댓글 및 답글 달기 구현

+++ app/src/main/java/com/byeboo/app/presentation/quest/QuestScreen.kt
@@ -154,7 +154,7 @@ private fun QuestScreen(
     onCommonAnswerClick: (Long) -> Unit,
     onDateChange: (LocalDate) -> Unit,
     onLoadMore: () -> Unit,
-    onHeartClick: () -> Unit,
+    onHeartClick: (Long) -> Unit,
 ) {
     if (uiState.myJourneyState.showQuitModal) {
         QuestModal(
